Rename: カブ価記録画面の名称を変更
