+++ app/pages/stock/form.tsx
@@ -2,7 +2,7 @@ import { NextPage } from 'next';
 import { connect } from 'react-redux';
 
 // component
-import Form from '@/components/page/Analytics/Form';
+import Form from '@/components/page/Stock/Form';
 
 const Page: NextPage = () => {
   return <Form title="カブ価を編集するフォーム" />;
