+++ app/style/page/stock.module.scss
@@ -1,6 +1,6 @@
 @import '../_variable';
 
-.p-analytics {
+.p-stock {
   .o-chart {
     position: relative;
     margin: 1rem 1rem 1rem 0;
